Write both jwts to storage
