+++ tee-worker/omni-executor/executor-storage/src/lib.rs
@@ -6,8 +6,8 @@ mod account_store;
 pub use account_store::AccountStoreStorage;
 mod oauth2_state_verifier;
 pub use oauth2_state_verifier::OAuth2StateVerifierStorage;
-mod pumpx_auth_token_id;
-pub use pumpx_auth_token_id::PumpxAuthTokenIdStorage;
+mod pumpx_jwt;
+pub use pumpx_jwt::PumpxJwtStorage;
 
 use executor_crypto::hashing::{blake2_128, twox_128};
 use executor_primitives::{AccountId, MemberAccount};
